Create README.md
ListKeeper is a simple sample android app written in Kotlin for creating, using, and maintaining lists.
It uses the Room Persistence Library, simple coroutines,  and is written as an MVVM architecture. ListKeeper uses ItemTouchHelper for its RecyclerViews allowing drag-and-drop reordering and swipe to delete. It also provides the ability to share a list and allows for speech-to-text data entry and editing.